+++ mlbusinesscomponents/src/main/java/com/mercadolibre/android/mlbusinesscomponents/components/touchpoint/view/carousel/CarouselView.java
@@ -15,6 +15,7 @@ import com.mercadolibre.android.mlbusinesscomponents.components.touchpoint.domai
 import com.mercadolibre.android.mlbusinesscomponents.components.touchpoint.domain.model.carousel.CarouselCard;
 import com.mercadolibre.android.mlbusinesscomponents.components.touchpoint.tracking.TouchpointTrackeable;
 import com.mercadolibre.android.mlbusinesscomponents.components.touchpoint.view.AbstractTouchpointChildView;
+import com.mercadolibre.android.mlbusinesscomponents.components.touchpoint.view.carousel.card.CarouselCardView;
 import com.mercadolibre.android.mlbusinesscomponents.components.touchpoint.view.carousel.card.TrackListener;
 import com.mercadolibre.android.mlbusinesscomponents.components.utils.ScaleUtils;
 import java.util.ArrayList;
@@ -33,6 +34,7 @@ public class CarouselView extends AbstractTouchpointChildView<Carousel> {
     private TrackListener trackListener;
     private TouchpointImageLoader imageLoader;
     private HorizontalScrollingEnhancer horizontalScrollingEnhancer;
+    private CarouselCardView templateCardView;
 
     /**
      * Constructor
@@ -66,6 +68,7 @@ public class CarouselView extends AbstractTouchpointChildView<Carousel> {
         recyclerView = findViewById(R.id.touchpoint_carousel_recycler_view);
         adapter = new CarouselAdapter();
         rect = new Rect();
+        templateCardView = findViewById(R.id.templateCardView);
         initList(context);
     }
 
@@ -133,8 +136,13 @@ public class CarouselView extends AbstractTouchpointChildView<Carousel> {
     }
 
     private void showCards(final List<CarouselCard> cards, final HeightCalculatorDelegate heightCalculator) {
-        adapter.setItems(cards);
-        adapter.setCardHeight(heightCalculator.getFixedCardHeight());
+        templateCardView.bind(cards.get(heightCalculator.getMaxHeightItemIndex()));
+        templateCardView.setVisibility(INVISIBLE);
+        templateCardView.post(() -> {
+            adapter.setCardHeight(templateCardView.getHeight());
+            adapter.setItems(cards);
+            templateCardView.setVisibility(GONE);
+        });
     }
 
     @Override