+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## NEW
+* FIX - Se corrige el calculo de la altura de los items touchpoint
 * FIX - Se corrige label croppeado al aumentar el tamaño del texto
 * FIX - Se corrige la imagen croppeada en hybrid y default Tochview